agm: Correct typecasting in compress plugin ops.
This change aligns the plugin ops structure with the expected format
used by TinyCompress.  Add missing pthread.h and string.h include to
resolve implicit declaration warnings

+++ plugins/tinyalsa/src/agm_compress_plugin.c
@@ -302,7 +302,7 @@ err:
     return ret;
 }
 
-int agm_compress_write(void *data, const void *buf, unsigned int size)
+int agm_compress_write(void *data, const void *buf, size_t size)
 {
     struct agm_compress_priv *priv = data;
     struct snd_compr_avail avail;
@@ -383,7 +383,7 @@ int agm_read(struct agm_compress_priv *priv, void *buff, size_t count)
     return count;
 }
 
-int agm_compress_read(void *data, void *buf, unsigned int size)
+int agm_compress_read(void *data, void *buf, size_t size)
 {
     struct agm_compress_priv *priv = data;
     struct snd_compr_avail avail;
@@ -458,7 +458,7 @@ int agm_compress_tstamp(void *data,
 }
 
 int agm_compress_get_tstamp(void *data,
-                        unsigned long *samples, unsigned int *sampling_rate)
+                        unsigned int *samples, unsigned int *sampling_rate)
 {
     struct agm_compress_priv *priv = data;
     struct snd_compr_tstamp ktstamp;
